services/surya-ocr/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.9.0
surya-ocr==0.17.1
torch>=2.9.0
opencv-python>=4.10.0
numpy>=2.0.0
pillow>=10.0.0
openpyxl>=3.1.5
xlrd>=2.0.1
pymupdf>=1.24.0
pdfplumber>=0.11.0
pandas>=2.2.0
